examples/requirements.txt
config_loader>=0.1.0
cryptography>=41.0.0
aiohttp>=3.8.0
pyyaml>=6.0.0 